Move Command to interfaces dir
