+++ src/interfaces/command.ts
@@ -1,4 +1,4 @@
-import {MsgContext} from '../interfaces';
+import {MsgContext} from './msg_context';
 
 export interface Command {
   matches(name: string): boolean;
