+++ src/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './msg_context';
+export * from './command';
