+++ test/bot/command_handler.test.ts
@@ -1,6 +1,5 @@
-import {Command} from '../../src/bot/command';
 import {CommandHandler} from '../../src/bot/command_handler';
-import {MsgContext} from '../../src/interfaces';
+import {MsgContext, Command} from '../../src/interfaces';
 import {FakeMsgContext} from '../fakes';
 
 const MockCommandMatcher = jest.fn<jest.Mocked<Command>, [string]>((cmdName: string) => ({
